project_07_real_time_od_websocket/requirements.txt
fastapi
uvicorn[standard]
httpie
transformers[torch]
Pillow
python-multipart